third_party/WebKit/LayoutTests/typedcssom/computedstyle/custom-properties-expected.txt

This is a testharness.js-based test.
FAIL Getting a length type registered property returns a CSSUnitValue assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Getting a URL image type registered property returns a CSSURLImageValue assert_equals: expected "loaded" but got "unloaded"
FAIL Getting a width with a length type var value returns a CSSUnitValue assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
PASS Getting a background-image with a URL image type var value returns a CSSURLImageValue 
FAIL Getting an inherited length type registered property returns a CSSUnitValue assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Getting a non-inherited length type registered property returns a CSSUnitValue assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Getting a height with a calc type containing var values returns a CSSCalcValue CSSCalcValue is not defined
FAIL Getting the value of a registered property that isn't set on the element returns the initial value for the property assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
PASS Getting the value of a property that isn't registered returns null 
FAIL getAll for a length type registered property returns a single value assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
Harness: the test ran to completion.
